Fix MSVC warnings

+++ src/optix_api.cpp
@@ -531,7 +531,7 @@ bool jitc_optix_compile(ThreadState *ts, const char *buf, size_t buf_size,
     }
 
     std::function<void(OptixTask)> execute_task = [&](OptixTask task) {
-        size_t max_new_tasks = pool_size();
+        unsigned int max_new_tasks = pool_size();
 
         std::unique_ptr<OptixTask[]> new_tasks =
             std::make_unique<OptixTask[]>(max_new_tasks);
